More

+++ app/src/main/java/com/tgarbus/posturecheck/NotificationResponseService.kt
@@ -17,20 +17,28 @@ class NotificationResponseService : Service() {
   private val job = SupervisorJob()
   private val scope = CoroutineScope(job)
 
-    override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
+  override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
     super.onStartCommand(intent, flags, startId)
     val context = this
-    val pastCheck: PastPostureCheck = PastPostureCheck.fromBundle(intent!!.extras!!)
+    val bundle = intent!!.extras!!
+    val notificationManager = getSystemService(NOTIFICATION_SERVICE) as NotificationManager
+    if (bundle.getBoolean("is_test_notification", false)) {
+      notificationManager.cancel(kTestNotificationId)
+      // TODO: Document why this is needed.
+      stopSelf(startId)
+      return START_REDELIVER_INTENT
+    }
+    val pastCheck: PastPostureCheck = PastPostureCheck.fromBundle(bundle)
     scope.launch {
       val plannedChecksRepo = PlannedChecksRepository(context)
       val pastChecksRepo = PastChecksRepository(context)
       val plannedCheck = pastCheck.withoutReply()
       pastChecksRepo.addPastCheck(pastCheck)
       plannedChecksRepo.deletePlannedCheck(plannedCheck)
+      notificationManager.cancel(pastCheck.notificationId())
+      // TODO: Document why this is needed.
       stopSelf(startId)
     }
-    val notificationManager = getSystemService(NOTIFICATION_SERVICE) as NotificationManager
-    notificationManager.cancel(pastCheck.notificationId())
     return START_REDELIVER_INTENT
   }
